requirements.txt
flask>=2.3.0
torch>=2.0.0
torchvision>=0.15.0
requests>=2.31.0
pillow>=9.0.0
